feat: Improve shadow-ability of project template
